Update README.md
In the main() function, add statements to output the following welcome message (the same as it appears here) to the console:

This program finds the average and
mean square of 2 numbers.

Then, add statements prompting the user to input two numbers from the keyboard which should be stored in two floating-point variables.

Next, add statements to calculate the average and mean square of the 2 numbers. To square a number you can multiply it by itself. The average is the sum of both numbers divided by 2.

Finally, add statements to output a message stating the average and mean square of the two inputs rounded to 2 decimal places. For example, for an input of 3.2 and 5.3 it could output:

The average is 4.25 and
the mean square is 19.16

Note: 19.17 would also okay for mean square.

Be sure to double check the answers from your program with the sample output.